requirements.txt
ws4py
python-cjson
